code/python/requirements.txt
build==0.4.0
chardet==4.0.0
importlib-metadata==4.3.0
nibabel==3.2.1
numpy==1.20.3
packaging==20.9
pep517==0.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
